requirements.txt
pandas
numpy
google-cloud-aiplatform
python-dotenv